Pcm Player simplified.
