--- src/main/java/org/runejs/client/sound/PcmPlayerBase.java
@@ -1,14 +1,0 @@
-package org.runejs.client.sound;
-
-public class PcmPlayerBase {
-    public PcmPlayerBase(int arg0) {
-        StaticAudio.sampleRate = arg0;
-        StaticAudio.aLong288 = System.currentTimeMillis();
-    }
-
-    public void method212(long arg0) {
-    }
-
-    public void method213() {
-    }
-}
